24/06/05

+++ src/main/java/com/example/mansshop_boot/domain/dto/main/MainListDTO.java
@@ -6,6 +6,6 @@ public record MainListDTO(
         String productId
         , String productName
         , String thumbnail
-        , long productPrice
+        , int productPrice
 ) {
 }
